bookmark controller/service done && e2e test completed

+++ src/user/user.service.ts
@@ -1,6 +1,7 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { EditUserDto } from './dto';
+import { use } from 'passport';
 
 @Injectable()
 export class UserService {
@@ -22,10 +23,18 @@ export class UserService {
     }
 
     async deleteUser(userId: number) {
-        const user = await this.prisma.user.delete({
+        const user = await this.prisma.user.findUnique({
             where: {
                 id: userId,
             }
         });
+        if (!user || user.id != userId) {
+            throw new ForbiddenException('Access to resources denied');
+        }
+        await this.prisma.user.delete({
+            where: {
+                id: userId,
+            }
+        })
     }
 }
